moved member-list into lead command
