--- locales/en-US/expoter.ftl
@@ -1,7 +1,0 @@
-command-name = export-roles
-command-description = Exports a list of the users in each role as a csv file
-
-role-option-name = target
-role-option-description = The role from with to get list
-
-message-followup = Members in {$role} role
