+++ src/commands/chat/execution/lead/region/role.ts
@@ -3,7 +3,7 @@ import { ChatInputCommandInteraction, Snowflake } from 'discord.js';
 import Logger from '../../../../../structures/Logger';
 import { memberState } from '../../../../../structures/helpers';
 import { t } from '../../../../../i18n';
-import { ns } from '../index';
+import { ns } from '../../../builders/lead';
 
 const regionLeadRoleID: Snowflake = process.env.REGIONAL_ROLE_ID;
